requirements.txt
pandas==3.0.1
numpy==2.4.3
matplotlib==3.10.8
seaborn==0.13.2
scikit-learn==1.8.0
dagshub==0.6.9
mlflow==3.10.1
pytest==9.0.2
python-dotenv==1.0.0